requirements.txt
streamlit
pandas
requests
beautifulsoup4
pdfplumber
python-dotenv
apscheduler
openpyxl
flask
flask-cors
